ML/PCA/Project_Dimensionality_Reduction/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
PyYAML>=6.0
azureml-core>=1.40.0
python-dotenv>=0.19.0
pytest>=7.0.0
